requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==10.0.1
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==6.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4